requirements.txt
biopython==1.8.0
higher==0.2.1
numpy==1.19.2
scipy==1.9.2
python-Levenshtein==0.12.0
pandas==1.5.0
pandas-datareader==0.9.0
pandoc==1.0.2
pandocfilters==1.4.2
scikit-learn==0.24.2
seaborn==0.11.1
matplotlib==3.4.3
tqdm==4.50.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
torch==1.11.0
torchmetrics==0.7.2
git+https://github.com/jwillis0720/ANT.git
